+++ Motor3D/Motor3D/Ponto.h
@@ -1,13 +1,13 @@
 #pragma once
 class Ponto {
-	double x;
-	double y;
-	double z;
+	float x;
+	float y;
+	float z;
 public:
-	Ponto(double,double,double);
-	double getX() { return x; }
-	double getY() { return y; }
-	double getZ() { return z; }
+	Ponto(float,float,float);
+	float getX() { return x; }
+	float getY() { return y; }
+	float getZ() { return z; }
 	virtual ~Ponto(void);
 };
